+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 - Update sandbox to allow caller to control emulator ports
 - Update `djangae.contrib.backup` and `djangae.contrib.locking` URLs to use the Django 2 `path` function
 - Fix backup task credentials for GAE Python 3
+- Fix `djangae.contrib.security.commands_utils.extract_views_from_urlpatterns` function
 
 
 ### Bug fixes: